refactor: make code more readable (#48)

+++ firmware/utils/custom_stdlib/custom_conversion.c
@@ -216,7 +216,7 @@ uint32_t convertFloat(float value, char* buffer) {
     (void)buffer;
 
     // Extract integer part
-    const int32_t ipart = (uint16_t)value;
+    const int32_t ipart = (int32_t)value;
 
     // Extract floating part
     float fpart = value - (float)ipart;
